+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 1.2.0]
+
+### Added
+
+- New 'check' command validates the table, shows a possible set arrangement and shows if any placed jokers can be freely used for other purposes without invalidating the table tile arrangement.
 
 ## [1.1.2] - 2021-03-25